Changed scalar channel datatype to fp64 for increased resolution. Implemented readback of sweeps as channel streams for RF signal generators and SSG5000X series

+++ scopehal/Stream.h
@@ -96,7 +96,7 @@ public:
 	WaveformBase* m_waveform;
 
 	///@brief The current value (only meaningful for analog scalar type)
-	float m_value;
+	double m_value;
 
 	///@brief General datatype stored in the stream
 	StreamType m_stype;
